+++ bloom_nofos/nofos/views.py
@@ -10,6 +10,7 @@ from django.conf import settings
 from django.contrib import messages
 from django.contrib.admin.views.decorators import staff_member_required
 from django.contrib.auth.decorators import user_passes_test
+from django.core.exceptions import PermissionDenied
 from django.db import transaction
 from django.db.models import F
 from django.http import HttpResponse, HttpResponseBadRequest
@@ -39,7 +40,11 @@ from .forms import (
     NofoThemeForm,
     SubsectionForm,
 )
-from .mixins import check_nofo_group_permission, GroupAccessMixin
+from .mixins import (
+    check_nofo_group_permission,
+    GroupAccessObjectMixin,
+    has_nofo_group_permission_func,
+)
 from .models import THEME_CHOICES, Nofo, Section, Subsection
 from .nofo import (
     add_body_if_no_body,
@@ -115,9 +120,21 @@ class NofosListView(ListView):
         return context
 
 
-class NofosDetailView(GroupAccessMixin, DetailView):
+class NofosDetailView(DetailView):
     model = Nofo
     template_name = "nofos/nofo_view.html"
+
+    def dispatch(self, request, *args, **kwargs):
+        nofo = get_object_or_404(Nofo, pk=kwargs.get("pk"))
+
+        # Most non-authed users are filtered out by middleware.py
+        if request.user.is_authenticated:
+            # do not let users from other groups print this nofo
+            if not has_nofo_group_permission_func(request.user, nofo):
+                raise PermissionDenied("You don’t have permission to view this NOFO.")
+
+        # Continue with the normal flow for anonymous or authorized users
+        return super().dispatch(request, *args, **kwargs)
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
@@ -146,7 +163,7 @@ class NofosDetailView(GroupAccessMixin, DetailView):
         return context
 
 
-class NofosEditView(GroupAccessMixin, DetailView):
+class NofosEditView(GroupAccessObjectMixin, DetailView):
     model = Nofo
     template_name = "nofos/nofo_edit.html"
 
@@ -160,7 +177,7 @@ class NofosEditView(GroupAccessMixin, DetailView):
         return context
 
 
-class NofosDeleteView(GroupAccessMixin, DeleteView):
+class NofosDeleteView(GroupAccessObjectMixin, DeleteView):
     model = Nofo
     success_url = reverse_lazy("nofos:nofo_index")
 
@@ -360,7 +377,7 @@ def nofo_import(request, pk=None):
         )
 
 
-class BaseNofoEditView(GroupAccessMixin, UpdateView):
+class BaseNofoEditView(GroupAccessObjectMixin, UpdateView):
     model = Nofo
 
     def get_success_url(self):
@@ -535,9 +552,11 @@ def nofo_subsection_edit(request, pk, subsection_pk):
     )
 
 
-class PrintNofoAsPDFView(GroupAccessMixin, View):
+class PrintNofoAsPDFView(GroupAccessObjectMixin, DetailView):
+    model = Nofo
+
     def post(self, request, pk):
-        nofo = get_object_or_404(Nofo, pk=pk)
+        nofo = self.get_object()
 
         # the absolute uri is points to the /edit page, so remove that from the path
         nofo_url = request.build_absolute_uri(nofo.get_absolute_url()).replace(
@@ -595,7 +614,7 @@ class PrintNofoAsPDFView(GroupAccessMixin, View):
             )
 
 
-class CheckNOFOLinksDetailView(GroupAccessMixin, DetailView):
+class CheckNOFOLinksDetailView(GroupAccessObjectMixin, DetailView):
     model = Nofo
     template_name = (
         "nofos/nofo_check_links.html"  # Replace with your actual template name